requirements.txt
# Core dependencies
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Google APIs (for Sheets)
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# Dropbox API
dropbox==11.36.2

# PDF generation
reportlab>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dateutil>=2.8.2
pandas>=2.0.0

